rename unk_021E7D8 to camera.c
